ASRC-142: properly validate EditMultiSelectVariable
EditVariableController needs to validate the EditExistingVar implementation
using the appropriate Validator for that implementation. To accomplish this, I
created a new interface, EditVar, that specifies getValidator().

Regrettably, there was already an abstract class EditVar which I renamed to
EditBaseVar to make way for the interface. This resulted in many accompanying
changes not directly related to fixing the validation.

EditBaseVar and EditExistingVar now extend the new EditVar interface.
