+++ srcalc/src/main/webapp/WEB-INF/views/fragments/multiSelectProperties.jspf
@@ -1,3 +1,4 @@
+<%@ taglib uri="/WEB-INF/srcalc.tld" prefix="srcalc" %>
 <%@ taglib uri="http://java.sun.com/jsp/jstl/core" prefix="c" %>
 <%@ taglib uri="http://www.springframework.org/tags/form" prefix="form" %>
 <%--
@@ -23,7 +24,15 @@ intended as the body of an <editVariablePage> tag.
             --%>
             <c:set var="i" value="0" />
             <c:forEach items="${variable.options}">
-            <li><form:input path="options[${i}]"/> <form:errors path="options[${i}]" cssClass="error"/></li>
+            <li>
+                <%--
+                Use the given optionLengthMax for the text box size, but cap it
+                at 40 because any bigger is too much longer than the expected
+                length.
+                 --%>
+                <form:input path="options[${i}]" size="${srcalc:min(variable.optionLengthMax, 40)}"/>
+                <form:errors path="options[${i}]" cssClass="error"/>
+            </li>
             <c:set var="i" value="${i+1}" />
             </c:forEach>
             </ol>
